Refactor memory view and use .scrollable()
